requirements.txt
crewai
flask
markdown
litellm
tavily-python